+++ src/MicroJsonComposer.h
@@ -24,7 +24,7 @@ public:
   void addProperty(const char *key, String value);
 
   // Create a property with int value
-  void addProperty(const char *key, int value);
+  void addProperty(const char *key, long value);
 
   const char *stringify();
